Fixes lights flickering from mirror camera
I think multiple cameras causes the vertices to move back and forth slightly when being rendered, which was causing the lights to flicker through them. I "fixed" it by just putting the lights on the other side of the wall. Since the walls are just planes some light still gets through, and I think it looks okay.